# swagger.yaml
openapi: 3.0.0
info:
  title: restaurants-api
  contact: {}
  version: '1.0'
servers:
- url: https://alejandro-restaurant-api.herokuapp.com/api/v1
  variables: {}
paths:
...
  /users/register:
    post:
      tags:
      - Auth
      summary: Register User
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
            example:
              name: Hector Harm
              email: Harm@gmail.com
              password: hola
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    delete:
      tags:
      - User
      summary: Delete User
      operationId: DeleteUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - User
      summary: Get One User
      operationId: GetOneUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    patch:
      tags:
      - User
      summary: Update User
      operationId: UpdateUser
      parameters:
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: hola?
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              text: never mind
              user_id: 61eb300ac817a8ab88a0731e
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
...
    RegisterUserRequest:
      title: RegisterUserRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: si?..
        email: alejandro1@gmail.com
        password: hola
...
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
      - text
      - user_id
      type: object
      properties:
        text:
          type: string
        user_id:
          type: string
      example:
        text: never mind
        user_id: 61eb300ac817a8ab88a0731e
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...